pyproject.toml
[tool.poetry]
name = "say-apiv3"
version = "0.1.0"
description = ""
authors = ["Arash Fatahzade <fatahzade@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
drf-spectacular = "^0.13.1"
djangorestframework-simplejwt = "^4.6.0"
psycopg2 = "^2.8.6"
celery = {extras = ["redis"], version = "^5.0.5"}
djangorestframework-camel-case = "^1.2.0"
Jinja2 = "^2.11.2"
pyotp = "^2.5.1"
django-redis = "^4.12.1"

[tool.poetry.dev-dependencies]
pudb = "^2020.1"
jedi = "^0.18.0"
pytest-pythonpath = "^0.7.3"
pytest-django = "^4.1.0"
Faker = "^6.4.1"
model_bakery = "^1.2.1"
pytest-cov = "^2.11.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
